+++ tests/pytests/pkg/integration/test_enabled_disabled.py
@@ -3,10 +3,12 @@ from pytestskipmarkers.utils import platform
 
 
 @pytest.mark.skip_on_windows(reason="Linux test only")
-def test_services(install_salt, salt_cli, salt_minion):
+def test_services(install_salt, salt_cli, salt_minion, salt_call_cli):
     """
     Check if Services are enabled/disabled
     """
+    install_salt.install()
+
     services_disabled = []
     services_enabled = []
     if install_salt.distro_id in ("ubuntu", "debian"):
@@ -29,9 +31,15 @@ def test_services(install_salt, salt_cli, salt_minion):
         pytest.fail(f"Don't know how to handle os_family={install_salt.distro_id}")
 
     for service in services_enabled:
-        ret = salt_cli.run("service.enabled", service, minion_tgt=salt_minion.id)
-        assert "true" in ret.stdout
+        test_cmd = f"systemctl show -p UnitFileState {service}"
+        ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
+        test_enabled = ret.stdout.strip().split("=")[1].split('"')[0].strip()
+        assert ret.returncode == 0
+        assert test_enabled == "enabled"
 
     for service in services_disabled:
-        ret = salt_cli.run("service.disabled", service, minion_tgt=salt_minion.id)
-        assert "true" in ret.stdout
+        test_cmd = f"systemctl show -p UnitFileState {service}"
+        ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
+        test_enabled = ret.stdout.strip().split("=")[1].split('"')[0].strip()
+        assert ret.returncode == 0
+        assert test_enabled == "disabled"
