+++ tests/pytests/pkg/upgrade/test_salt_upgrade.py
@@ -1,12 +1,9 @@
-import logging
 import time
 
 import packaging.version
 import psutil
 import pytest
 from pytestskipmarkers.utils import platform
-
-log = logging.getLogger(__name__)
 
 pytestmark = [pytest.mark.skip_unless_on_linux(reason="Only supported on Linux family")]
 
@@ -145,295 +142,3 @@ def test_salt_upgrade(salt_call_cli, install_salt):
             # test pip install after an upgrade
             use_lib = salt_call_cli.run("--local", "github.get_repo_info", repo)
             assert "Authentication information could" in use_lib.stderr
-
-
-def test_salt_systemd_disabled_preservation(
-    salt_call_cli, install_salt, salt_systemd_setup
-):
-    """
-    Test upgrade of Salt packages preserve disabled state of systemd
-    """
-    if not install_salt.upgrade:
-        pytest.skip("Not testing an upgrade, do not run")
-
-    if install_salt.relenv:
-        original_py_version = install_salt.package_python_version()
-
-    # setup systemd to enabled and active for Salt packages
-    # pylint: disable=pointless-statement
-    salt_systemd_setup
-
-    # ensure known state, disabled
-    test_list = ["salt-api", "salt-minion", "salt-master"]
-    for test_item in test_list:
-        test_cmd = f"systemctl disable {test_item}"
-        ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-        assert ret.returncode == 0
-
-    # perform Salt package upgrade test
-    # pylint: disable=pointless-statement
-    salt_test_upgrade
-
-    # test for disabled systemd state
-    test_list = ["salt-api", "salt-minion", "salt-master"]
-    for test_item in test_list:
-        test_cmd = f"systemctl show -p UnitFileState {test_item}"
-        ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-        test_enabled = ret.stdout.strip().split("=")[1].split('"')[0].strip()
-        assert ret.returncode == 0
-        assert test_enabled == "disabled"
-
-
-def test_salt_systemd_enabled_preservation(
-    salt_call_cli, install_salt, salt_systemd_setup
-):
-    """
-    Test upgrade of Salt packages preserve enabled state of systemd
-    """
-    if not install_salt.upgrade:
-        pytest.skip("Not testing an upgrade, do not run")
-
-    if install_salt.relenv:
-        original_py_version = install_salt.package_python_version()
-
-    # setup systemd to enabled and active for Salt packages
-    # pylint: disable=pointless-statement
-    salt_systemd_setup
-
-    # perform Salt package upgrade test
-    # pylint: disable=pointless-statement
-    salt_test_upgrade
-
-    # test for enabled systemd state
-    test_list = ["salt-api", "salt-minion", "salt-master"]
-    for test_item in test_list:
-        test_cmd = f"systemctl show -p UnitFileState {test_item}"
-        ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-        test_enabled = ret.stdout.strip().split("=")[1].split('"')[0].strip()
-        assert ret.returncode == 0
-        assert test_enabled == "enabled"
-
-
-def test_salt_systemd_inactive_preservation(
-    salt_call_cli, install_salt, salt_systemd_setup
-):
-    """
-    Test upgrade of Salt packages preserve inactive state of systemd
-    """
-    if not install_salt.upgrade:
-        pytest.skip("Not testing an upgrade, do not run")
-
-    if install_salt.relenv:
-        original_py_version = install_salt.package_python_version()
-
-    # setup systemd to enabled and active for Salt packages
-    # pylint: disable=pointless-statement
-    salt_systemd_setup
-
-    # ensure known state, disabled
-    test_list = ["salt-api", "salt-minion", "salt-master"]
-    for test_item in test_list:
-        test_cmd = f"systemctl stop {test_item}"
-        ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-        assert ret.returncode == 0
-
-    # perform Salt package upgrade test
-    # pylint: disable=pointless-statement
-    salt_test_upgrade
-
-    # test for inactive systemd state
-    test_list = ["salt-api", "salt-minion", "salt-master"]
-    for test_item in test_list:
-        test_cmd = f"systemctl is-active {test_item}"
-        ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-        test_active = ret.stdout.strip().split()[2].strip('"').strip()
-        assert ret.returncode == 1
-        assert test_active == "inactive"
-
-
-def test_salt_systemd_active_preservation(
-    salt_call_cli, install_salt, salt_systemd_setup
-):
-    """
-    Test upgrade of Salt packages preserve active state of systemd
-    """
-    if not install_salt.upgrade:
-        pytest.skip("Not testing an upgrade, do not run")
-
-    if install_salt.relenv:
-        original_py_version = install_salt.package_python_version()
-
-    # setup systemd to enabled and active for Salt packages
-    # pylint: disable=pointless-statement
-    salt_systemd_setup
-
-    # perform Salt package upgrade test
-    # pylint: disable=pointless-statement
-    salt_test_upgrade
-
-    # test for active systemd state
-    test_list = ["salt-api", "salt-minion", "salt-master"]
-    for test_item in test_list:
-        test_cmd = f"systemctl is-active {test_item}"
-        ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-        test_active = ret.stdout.strip().split()[2].strip('"').strip()
-        assert ret.returncode == 0
-        assert test_active == "active"
-
-
-def test_salt_ownership_permission(salt_call_cli, install_salt, salt_systemd_setup):
-    """
-    Test upgrade of Salt packages preserve existing ownership
-    """
-    if not install_salt.upgrade:
-        pytest.skip("Not testing an upgrade, do not run")
-
-    if install_salt.relenv:
-        original_py_version = install_salt.package_python_version()
-
-    # setup systemd to enabled and active for Salt packages
-    # pylint: disable=pointless-statement
-    salt_systemd_setup
-
-    # test ownership for Minion, Master and Api
-    test_list = ["salt-api", "salt-minion", "salt-master"]
-    for test_item in test_list:
-        if "salt-api" == test_item:
-            test_cmd = f"ls -dl /run/{test_item}.pid"
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_user = ret.stdout.strip().split()[4]
-            assert ret.returncode == 0
-            assert test_user == "salt"
-
-            test_cmd = f"ls -dl /run/{test_item}.pid"
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_group = ret.stdout.strip().split()[5]
-            assert ret.returncode == 0
-            assert test_group == "salt"
-        else:
-            test_name = test_item.strip().split("-")[1]
-            test_cmd = f"ls -dl /run/salt/{test_name}"
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_user = ret.stdout.strip().split()[4]
-            assert ret.returncode == 0
-            if test_item == "salt-minion":
-                assert test_user == "root"
-            else:
-                assert test_user == "salt"
-
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_group = ret.stdout.strip().split()[5]
-            assert ret.returncode == 0
-            if test_item == "salt-minion":
-                assert test_group == "root"
-            else:
-                assert test_group == "salt"
-
-    # create master user, and minion user, change conf, restart and test ownership
-    test_master_user = "horse"
-    test_minion_user = "donkey"
-    ret = salt_call_cli.run("--local", "user.list_users")
-    user_list = ret.stdout.strip().split(":")[1]
-
-    if test_master_user not in user_list:
-        ret = salt_call_cli.run("--local", "user.add", f"{test_master_user}")
-
-    if test_minion_user not in user_list:
-        ret = salt_call_cli.run("--local", "user.add", f"{test_minion_user}")
-
-    ret = salt_call_cli.run(
-        "--local", "file.comment_line", "/etc/salt/master", "^user:"
-    )
-    assert ret.returncode == 0
-
-    ret = salt_call_cli.run(
-        "--local", "file.comment_line", "/etc/salt/minion", "^user:"
-    )
-    assert ret.returncode == 0
-
-    test_string = f"\nuser: {test_master_user}\n"
-    ret = salt_call_cli.run("--local", "file.append", "/etc/salt/master", test_string)
-
-    test_string = f"\nuser: {test_minion_user}\n"
-    ret = salt_call_cli.run("--local", "file.append", "/etc/salt/minion", test_string)
-
-    # restart and check ownership is correct
-    test_list = ["salt-api", "salt-minion", "salt-master"]
-    for test_item in test_list:
-        test_cmd = f"systemctl restart {test_item}"
-        ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-        assert ret.returncode == 0
-
-    time.sleep(10)  # allow some time for restart
-
-    # test ownership for Minion, Master and Api - horse and donkey
-    test_list = ["salt-api", "salt-minion", "salt-master"]
-    for test_item in test_list:
-        if "salt-api" == test_item:
-            test_cmd = f"ls -dl /run/{test_item}.pid"
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_user = ret.stdout.strip().split()[4]
-            assert ret.returncode == 0
-            assert test_user == f"{test_master_user}"
-
-            test_cmd = f"ls -dl /run/{test_item}.pid"
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_group = ret.stdout.strip().split()[5]
-            assert ret.returncode == 0
-            assert test_group == f"{test_master_user}"
-        else:
-            test_name = test_item.strip().split("-")[1]
-            test_cmd = f"ls -dl /run/salt/{test_name}"
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_user = ret.stdout.strip().split()[4]
-            assert ret.returncode == 0
-            if test_item == "salt-minion":
-                assert test_user == f"{test_minion_user}"
-            else:
-                assert test_user == f"{test_master_user}"
-
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_group = ret.stdout.strip().split()[5]
-            assert ret.returncode == 0
-            if test_item == "salt-minion":
-                assert test_group == f"{test_minion_user}"
-            else:
-                assert test_group == f"{test_master_user}"
-
-    # perform Salt package upgrade test
-    # pylint: disable=pointless-statement
-    salt_test_upgrade
-
-    # test ownership for Minion, Master and Api
-    test_list = ["salt-api", "salt-minion", "salt-master"]
-    for test_item in test_list:
-        if "salt-api" == test_item:
-            test_cmd = f"ls -dl /run/{test_item}.pid"
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_user = ret.stdout.strip().split()[4]
-            assert ret.returncode == 0
-            assert test_user == f"{test_master_user}"
-
-            test_cmd = f"ls -dl /run/{test_item}.pid"
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_group = ret.stdout.strip().split()[5]
-            assert ret.returncode == 0
-            assert test_group == f"{test_master_user}"
-        else:
-            test_name = test_item.strip().split("-")[1]
-            test_cmd = f"ls -dl /run/salt/{test_name}"
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_user = ret.stdout.strip().split()[4]
-            assert ret.returncode == 0
-            if test_item == "salt-minion":
-                assert test_user == f"{test_minion_user}"
-            else:
-                assert test_user == f"{test_master_user}"
-
-            ret = salt_call_cli.run("--local", "cmd.run", test_cmd)
-            test_group = ret.stdout.strip().split()[5]
-            assert ret.returncode == 0
-            if test_item == "salt-minion":
-                assert test_group == f"{test_minion_user}"
-            else:
-                assert test_group == f"{test_master_user}"
